refactor: consolidate web collab runtime files
